Get rid of more event_names
