+++ posthog/api/user.py
@@ -266,11 +266,6 @@ def user(request):
                 "api_token": team.api_token,
                 "anonymize_ips": team.anonymize_ips,
                 "slack_incoming_webhook": team.slack_incoming_webhook,
-                "event_names": team.event_names,
-                "event_names_with_usage": team.get_latest_event_names_with_usage(),
-                "event_properties": team.event_properties,
-                "event_properties_numerical": team.event_properties_numerical,
-                "event_properties_with_usage": team.get_latest_event_properties_with_usage(),
                 "completed_snippet_onboarding": team.completed_snippet_onboarding,
                 "session_recording_opt_in": team.session_recording_opt_in,
                 "session_recording_retention_period_days": team.session_recording_retention_period_days,
